requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pyinstaller==6.2.0
pywin32==306
pyautogui==0.9.54
Pillow==10.2.0
